refactor: Refactor directories
